+++ test/core/cast.v
@@ -32,10 +32,6 @@ module top;
         $display("%b", 32'd4);
         $display("%b", 33'd4);
         $display("%b", 33'h1_FFFF_FFFF);
-        $display("%b", 32'd0);
-        $display("%b", 33'd4294967296);
-        $display("%b", 32'd1);
-        $display("%b", 33'd4294967297);
     end
 
     localparam [0:0] foo = 0;