+++ test/core/sign_cast.sv
@@ -1,7 +1,9 @@
 module top;
     initial begin
-        $display(signed'(4294967295));
-        $display(unsigned'(4294967295));
+        $display(signed'(1'b1));
+        $display(unsigned'(1'sb1));
+        $display(signed'(1));
+        $display(unsigned'(1));
         $display(signed'(-1));
         $display(unsigned'(-1));
     end